requirements.txt
torch
torchvision
timm
onnx
onnxruntime
numpy
pillow
tensorboard
pyyaml
matplotlib
#dev
ruff
pytest
#db
psycopg2
sqlalchemy
#api
fastapi
uvicorn
python-multipart